iot/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
requests
pydantic
python-dotenv
scikit-learn
pandas
joblib
twilio
python-multipart
geopy
httpx
python-jose[cryptography]
passlib[bcrypt]
email-validator
groq
gunicorn
